Source attribution while adjusting for sampling bias

Source attribution while adjusting for sampling bias

source.attribution.mcmc(dobs, dprior, control = list(seed = 42, mcmc.n =
  1000, verbose = 1, outfile = "SAMCMCv190327.rda"))

Arguments

dobs
Data.table of observed number of transmission events for each transmission pair category.

dprior
Data.table of draws from the prior distribution of sampling probabilities.

control
List of input arguments that control the behaviour of the source attribution algorithm:

"seed"Random number seed, for reproducibility.

"mcmc.n"Guide on the number of MCMC iterations. The actual number of iterations will be slightly larger, and a multiple of the number of iterations needed to complete one MCMC sweep through all free parameters.

"verbose"Flag to switch on/off verbose output.

"outfile"Full path name to which the MCMC output is saved to in RDA format, in an object called mc.

Value

If `outfile` is not specified, phyloflows MCMC output is returned, which is a list object. If `outfile` is specified, phyloflows MCMC output is written to an RDA file.

See also

Examples

require(data.table)

#> Loading required package: data.table

#> Warning: package ‘data.table’ was built under R version 3.5.2

require(phyloflows)
# load input data set
data(twoGroupFlows1, package="phyloflows")
dobs <- twoGroupFlows1$dobs
dprior <- twoGroupFlows1$dprior

#
# run MCMC and return output to R
# this is usually fine for simple source attribution models
# that don t take long to run
#
# NOT RUN {
control <- list(seed=42,    # seed for reproducibility
                mcmc.n=5e4, # guide on the number of MCMC iterations, the actual number of MCMC iterations may be a little bit larger
                verbose=0   # switch on for output to each MCMC iteration
                )
mc <- phyloflows:::source.attribution.mcmc(dobs, dprior, control)
# }

#
# run MCMC and save output to file
# this is recommended when the source attribution is more complex
# so you can deploy the MCMC on a high performance cluster
#

# NOT RUN {
mcmc.file <- 'mcmcm_output.rda'
control <- list(seed=42,    # seed for reproducibility
                mcmc.n=5e4, # guide on the number of MCMC iterations, the actual number of MCMC iterations may be a little bit larger
                verbose=0   # switch on for output to each MCMC iteration
                outfile= mcmc.file # store MCMC output as rda file here
                )
mc <- phyloflows:::source.attribution.mcmc(dobs, dprior, control)
# }