MCMC diagnostics for the source attribution algorithm

MCMC diagnostics for the source attribution algorithm

source.attribution.mcmc.diagnostics(mcmc.file, mc = NULL,
  control = list(burnin.p = 0.2, regex_pars = "*", credibility.interval =
  0.95, pdf.plot.n.worst.case.parameters = 10, pdf.plot.all.parameters =
  FALSE, pdf.height.per.par = 1.2, outfile.base = gsub("\\.rda", "",
  mcmc.file)))

Arguments

mcmc.file
Full file name to MCMC output from function source.attribution.mcmc

mc
MCMC object from function source.attribution.mcmc; this can be specified alternatively to `mcmc.file`.

control
List of input arguments that control the behaviour of the source attribution algorithm:

"burnin.p"Proportion of MCMC iterations that are removed as burn-in period.

"regex_pars"Regular expression to select the parameter names for which diagnostics are computed. The default is all parameters, which is '*'.

"credibility.interval"Width of the marginal posterior credibility intervals.

"pdf.plot.n.worst.case.parameters"Integer which specifies the number of parameters with smallest effective sample size that are inspected in detail. If set to 0, worst case analyses are not performed.

"pdf.plot.all.parameters"Flag which specifies if traces shall be plotted for all parameters. If set to TRUE, very large pdfs may be created.

"pdf.height.per.par"Most plots show diagnostics with parameters listed on the y-axis. This value controls the plot height in inches for each free parameter.

"outfile.base"Start of the full file name for all output files.

Value

Nothing. Diagnostic plots and csv files are written to disk.