Estimate Flows, Sources, WAIFM, Flow ratios

Estimate Flows, Sources, WAIFM, Flow ratios

source.attribution.mcmc.getKeyQuantities(infile = NULL, mc = NULL,
  pars = NULL, dobs = NULL, control = NULL)

Arguments

infile
Full file name to MCMC output or aggregated MCMC output from function source.attribution.mcmc

control
List of input arguments that control the behaviour of the derivation of key quantities:

"quantiles"Named list of quantiles. Default: c('CL'=0.025,'IL'=0.25,'M'=0.5,'IU'=0.75,'CU'=0.975)

"flowratios"Cector of length 3. First element: name of flow ratio. Second element: name of transmission pair category for enumerator of flow ratio. Third element: name of transmission pair category for denominator of flow ratio.

"outfile"Full file name for output csv file.

Value

If `outfile` is not specified, the output is returned as a data.table. If `outfile` is specified, the output is written to a csv file.

See also

Examples

require(data.table)
require(phyloflows)
# load input data set
data(twoGroupFlows1, package="phyloflows")
dobs <- twoGroupFlows1$dobs
dprior <- twoGroupFlows1$dprior
# load MCMC output obtained with
# control <- list(seed=42, mcmc.n=5e4, verbose=0)
# mc <- phyloflows:::source.attribution.mcmc(dobs, dprior, control)
data(twoGroupFlows1_mcmc, package="phyloflows")

#
# calculate sources, onward flows & flow ratios from MCMC output 
#
control <- list( quantiles= c('CL'=0.025,'IL'=0.25,'M'=0.5,'IU'=0.75,'CU'=0.975),
                 flowratios= list( c('1/2', '1 2', '2 1'), c('2/1', '2 1', '1 2'))
                 )
ans <- phyloflows:::source.attribution.mcmc.getKeyQuantities(mc=mc, dobs=dobs, control=control)

#> 
#> Computing flows...
#> Computing WAIFM...
#> Computing sources...
#> Computing flow ratios...

#> Warning: Adding new column '2 1' then assigning NULL (deleting it).

#> Warning: Adding new column '1 2' then assigning NULL (deleting it).

#
# calculate sources, onward flows & flow ratios from aggregated MCMC output 
#
daggregateTo <- subset(dobs, select=c(TRM_CAT_PAIR_ID, TR_TRM_CATEGORY, REC_TRM_CATEGORY))
daggregateTo[, TR_TARGETCAT:= TR_TRM_CATEGORY]

#>    TRM_CAT_PAIR_ID TR_TRM_CATEGORY REC_TRM_CATEGORY TR_TARGETCAT
#> 1:               1               1                1            1
#> 2:               2               1                2            1
#> 3:               3               2                1            2
#> 4:               4               2                2            2

daggregateTo[, REC_TARGETCAT:= 'Any']

#>    TRM_CAT_PAIR_ID TR_TRM_CATEGORY REC_TRM_CATEGORY TR_TARGETCAT REC_TARGETCAT
#> 1:               1               1                1            1           Any
#> 2:               2               1                2            1           Any
#> 3:               3               2                1            2           Any
#> 4:               4               2                2            2           Any

set(daggregateTo, NULL, c('TR_TRM_CATEGORY','REC_TRM_CATEGORY'), NULL)
control <- list( burnin.p=0.05, thin=NA_integer_, regex_pars='PI')
pars <- phyloflows:::source.attribution.mcmc.aggregateToTarget(mc=mc, daggregateTo=daggregateTo, control=control)

#> 
#> Using MCMC output specified as input...
#> Collecting parameters...
#> Removing burnin in set to  5 % of chain, total iterations= 500
#> Making aggregated MCMC output...
# NOT RUN {
control <- list( quantiles= c('CL'=0.025,'IL'=0.25,'M'=0.5,'IU'=0.75,'CU'=0.975),
                 flowratios= list( c('1/2', '1 2', '2 1'), c('2/1', '2 1', '1 2'))
                 )
ans <- phyloflows:::source.attribution.mcmc.getKeyQuantities(pars=pars, control=control)
# }

#
# calculate sources, onward flows & flow ratios from file 
#

# NOT RUN {
infile<- 'blah.rda'     # either MCMC output or aggregated MCMC output
control <- list( quantiles= c('CL'=0.025,'IL'=0.25,'M'=0.5,'IU'=0.75,'CU'=0.975),
                 flowratios= list( c('1/2', '1 2', '2 1'), c('2/1', '2 1', '1 2'))
                 )
ans <- phyloflows:::source.attribution.mcmc.getKeyQuantities(infile=infile, dobs=dobs, control=control)
# }