Source attribution while adjusting for sampling bias

Source attribution while adjusting for sampling bias

source.attribution.mcmc(dobs, dprior, control = list(seed = 42, mcmc.n =
  1000, verbose = 1, outfile = "SAMCMCv190327.rda"))

Arguments

dobs
Data.table of observed number of transmission events for each transmission pair category.

dprior
Data.table of draws from the prior distribution of sampling probabilities.

control
List of input arguments that control the behaviour of the source attribution algorithm:

"seed"Random number seed, for reproducibility.

"mcmc.n"Guide on the number of MCMC iterations. The actual number of iterations will be slightly larger, and a multiple of the number of iterations needed to complete one MCMC sweep through all free parameters.

"verbose"Flag to switch on/off verbose output.

"outfile"Full path name to which the MCMC output is saved to in RDA format, in an object called mc.

Value

NULL. MCMC output is written to an RDA file.